chore: simplify mcp classes even more
