+++ openhands/agenthub/codeact_agent/function_calling.py
@@ -42,7 +42,7 @@ from openhands.events.action.mcp import McpAction
 from openhands.events.event import FileEditSource, FileReadSource
 from openhands.events.tool import ToolCallMetadata
 from openhands.llm import LLM
-from openhands.mcp.mcp_base import BaseTool
+from openhands.mcp.mcp import BaseTool
 
 
 def combine_thought(action: Action, thought: str) -> Action:
