+++ openhands/mcp/__init__.py
@@ -1,0 +1,3 @@
+from openhands.mcp.mcp import MCPClient, convert_mcp_clients_to_tools, create_mcp_clients, BaseTool
+
+__all__ = ['MCPClient', 'convert_mcp_clients_to_tools', 'create_mcp_clients', 'BaseTool']
