--- openhands/mcp/mcp_base.py
@@ -1,103 +1,0 @@
-import json
-from abc import ABC, abstractmethod
-from typing import Any, Dict, Optional
-
-from mcp.types import ImageContent
-from pydantic import BaseModel, Field
-from litellm import ChatCompletionToolParam, ChatCompletionToolParamFunctionChunk
-
-
-class ExtendedImageContent(ImageContent):
-    url: str
-
-
-class BaseTool(ABC, BaseModel):
-    name: str
-    description: str
-    parameters: Optional[dict] = None
-
-    @classmethod
-    def postfix(cls) -> str:
-        return '_mcp_tool_call'
-
-    class Config:
-        arbitrary_types_allowed = True
-
-    async def __call__(self, **kwargs) -> Any:
-        """Execute the tool with given parameters."""
-        return await self.execute(**kwargs)
-
-    @abstractmethod
-    async def execute(self, **kwargs) -> Any:
-        """Execute the tool with given parameters."""
-
-    def to_param(self) -> Dict:
-        """Convert tool to function call format."""
-        return {
-            'type': 'function',
-            'function': {
-                'name': self.name + self.postfix(),
-                'description': self.description,
-                'parameters': self.parameters,
-            },
-        }
-
-
-class ToolResult(BaseModel):
-    """Represents the result of a tool execution."""
-
-    output: Any = Field(default=None)
-    error: Optional[str] = Field(default=None)
-    base64_image: Optional[str] = Field(default=None)
-    system: Optional[str] = Field(default=None)
-
-    class Config:
-        arbitrary_types_allowed = True
-
-    def __bool__(self):
-        return any(getattr(self, field) for field in self.__fields__)
-
-    def __add__(self, other: 'ToolResult'):
-        def combine_fields(
-            field: Optional[str], other_field: Optional[str], concatenate: bool = True
-        ):
-            if field and other_field:
-                if concatenate:
-                    return field + other_field
-                raise ValueError('Cannot combine tool results')
-            return field or other_field
-
-        return ToolResult(
-            output=combine_fields(self.output, other.output),
-            error=combine_fields(self.error, other.error),
-            base64_image=combine_fields(self.base64_image, other.base64_image, False),
-            system=combine_fields(self.system, other.system),
-        )
-
-    def __str__(self):
-        if self.error:
-            return f'Error: {self.error}'
-        elif isinstance(self.output, ExtendedImageContent):
-            return json.dumps(
-                {
-                    'type': 'image',
-                    'image_url': f'data:image/png;base64,{self.output.data}',
-                    'mimeType': self.output.mimeType,
-                    'url': self.output.url,
-                }
-            )
-        elif isinstance(self.output, ImageContent):
-            return json.dumps(
-                {
-                    'type': 'image',
-                    'image_url': f'data:image/png;base64,{self.output.data}',
-                    'mimeType': self.output.mimeType,
-                }
-            )
-        else:
-            return self.output
-
-    def replace(self, **kwargs):
-        """Returns a new ToolResult with the given fields replaced."""
-        # return self.copy(update=kwargs)
-        return type(self)(**{**self.dict(), **kwargs})
